+++ modules/dxp/apps/osb/osb-customer/osb-portlet/docroot/WEB-INF/src/com/liferay/osb/support/util/SupportUtil.java
@@ -946,7 +946,7 @@ public class SupportUtil {
 			}
 
 			if (listTypeName.equals(name)) {
-				return listType.getListTypeId();
+				return (int)listType.getListTypeId();
 			}
 		}
